Reuse deleteSLO application service

+++ x-pack/solutions/observability/plugins/slo/server/services/delete_slo.ts
@@ -14,10 +14,14 @@ import {
   getSLOTransformId,
   getWildcardPipelineId,
 } from '../../common/constants';
-import { SLODefinition } from '../domain/models';
 import { retryTransientEsErrors } from '../utils/retry';
 import { SLORepository } from './slo_repository';
 import { TransformManager } from './transform_manager';
+
+interface Options {
+  skipDataDeletion: boolean;
+  skipRuleDeletion: boolean;
+}
 
 export class DeleteSLO {
   constructor(
@@ -28,13 +32,22 @@ export class DeleteSLO {
     private rulesClient: RulesClientApi
   ) {}
 
-  public async execute(sloId: string): Promise<void> {
+  public async execute(
+    sloId: string,
+    options: Options = {
+      skipDataDeletion: false,
+      skipRuleDeletion: false,
+    }
+  ): Promise<void> {
     const slo = await this.repository.findById(sloId);
 
     // First delete the linked resources before deleting the data
+    const rollupTransformId = getSLOTransformId(slo.id, slo.revision);
+    const summaryTransformId = getSLOSummaryTransformId(slo.id, slo.revision);
+
     await Promise.all([
-      this.deleteSummaryTransform(slo),
-      this.deleteRollupTransform(slo),
+      this.transformManager.uninstall(rollupTransformId),
+      this.summaryTransformManager.uninstall(summaryTransformId),
       retryTransientEsErrors(() =>
         this.scopedClusterClient.asSecondaryAuthUser.ingest.deletePipeline(
           { id: getWildcardPipelineId(slo.id, slo.revision) },
@@ -45,23 +58,17 @@ export class DeleteSLO {
     ]);
 
     await Promise.all([
-      this.deleteRollupData(slo.id),
-      this.deleteSummaryData(slo.id),
-      this.deleteAssociatedRules(slo.id),
+      this.deleteRollupData(slo.id, options.skipDataDeletion),
+      this.deleteSummaryData(slo.id, options.skipDataDeletion),
+      this.deleteAssociatedRules(slo.id, options.skipRuleDeletion),
     ]);
   }
 
-  private async deleteRollupTransform(slo: SLODefinition) {
-    const rollupTransformId = getSLOTransformId(slo.id, slo.revision);
-    await this.transformManager.uninstall(rollupTransformId);
-  }
+  private async deleteRollupData(sloId: string, skip: boolean): Promise<void> {
+    if (skip) {
+      return;
+    }
 
-  private async deleteSummaryTransform(slo: SLODefinition) {
-    const summaryTransformId = getSLOSummaryTransformId(slo.id, slo.revision);
-    await this.summaryTransformManager.uninstall(summaryTransformId);
-  }
-
-  private async deleteRollupData(sloId: string): Promise<void> {
     await this.scopedClusterClient.asCurrentUser.deleteByQuery({
       index: SLI_DESTINATION_INDEX_PATTERN,
       wait_for_completion: false,
@@ -79,7 +86,11 @@ export class DeleteSLO {
     });
   }
 
-  private async deleteSummaryData(sloId: string): Promise<void> {
+  private async deleteSummaryData(sloId: string, skip: boolean): Promise<void> {
+    if (skip) {
+      return;
+    }
+
     await this.scopedClusterClient.asCurrentUser.deleteByQuery({
       index: SUMMARY_DESTINATION_INDEX_PATTERN,
       refresh: false,
@@ -97,13 +108,17 @@ export class DeleteSLO {
       },
     });
   }
-  private async deleteAssociatedRules(sloId: string): Promise<void> {
+  private async deleteAssociatedRules(sloId: string, skip: boolean): Promise<void> {
+    if (skip) {
+      return;
+    }
+
     try {
       await this.rulesClient.bulkDeleteRules({
         filter: `alert.attributes.params.sloId:${sloId}`,
       });
     } catch (err) {
-      // no-op: bulkDeleteRules throws if no rules are found.
+      // no-op
     }
   }
 }
